requirements.txt
streamlit
openai
python-dotenv
notion-client>=2.0.0
requests
httpx
beautifulsoup4
PyPDF2
# yt-dlp
edge-tts
pydub
audioop-lts
chromadb
pysqlite3-binary
sentence-transformers
loguru
langgraph
langchain-core